refactor(plugin): expose plugin
